Remove boolean return for xfce_notify_daemon_get_workarea.
It already has a fallback to get the monitor geometry, no need to do the
check twice. This fixes a weird bug were we end up with the monitor
geometry instead of the workarea and were xfce4-notifyd tries to place
the notifications over the panel which results in miss-aligned
notifications.