Delay the initialization of the placement data.
We compute the workareas and create the arrays when when show the first
notification. This fixes a bug where the workarea was not computed
correctly because it did not take the panels into account (because
gdk_screen_get_window_stack returns NULL during the first seconds of the
session).

It also makes startup a little faster and saves some memory
when the daemon is not needed and just queried for its capacities.