Fix close timeout logic (bug #7467).
The timeout was initialized every time a new notification was displayed
which makes no sense: now the timeout is initialized when all
notifications expire.

The tiemout is also removed when a new notification is shown and reset
to its initial value when querying server information/specifications
with no displayed notification.