Correctly updating the theme for notification windows, setting the widget class to 'app-notification' and 'osd' so current user theme is selected.
Always use a small rounded corners. This should not be necessary in
theory, as Adwaita defined border-radius: 0 0 6px 6px; for the
app-notification and osd css classes. The problem is that Gtk
for some reason gets the border-radius as gint and not as GtkBorder.
Getting this way the first value only, which is 0.