Don't kill the daemon after 10mins of inactivity (Bug #12754)
It is not clear anymore why this feature was ever useful
and I have hence chosen to drop it. It does not really
seem to save resources - in the worst case it takes some
time for the deamon to be started and the first notification
after a timeout is shown with a delay.