Fix support for hyperlinks in notifications (Bug #14073)
Try to set the body with markup and in case this fails (empty label)
fall back to escaping the whole string and showing it plainly.
This equals pango_parse_markup extended by checking for valid hyperlinks
which is not supported by pango).
This could be avoided if gtk_label_set_markup were gboolean instead of void.