log: properly handle iso8601 timestamps
The panel-plugin menu list and the config log menu list both depend on
iso8601 timestamps to correctly display the timestamp in the item tooltip.
However, a timezone was not specified, resulting in a time of "NULL"
being displayed.

This change modifies common/xfce-notify-log to write out iso8601
timestamps directly using the is8601  g_date_time* functions.
It then adds timezone localization support for the tooltips.

This change also fixes some memory leaks handling the dates
in both the config menu and the panel plugin menu.