settings: Show 'Unspecified app' instead of empty line
As applications don't have to provide their name when sending
notifications we get an empty line in the list of known applications. In
theory we could just hide this line, but by keeping it around we enable
users to mute unspecified applications.
By setting the default string we expose this feature in a more user
friendly and understandable way.