settings: Better application icon matching
Leverage GIO's g_desktop_app_info_search function to fuzzy-find a match
for the application if we otherwise couldn't find a match. We naively
pick the first match because the matching algorithm is unknown and
potentially subject to change according to the GIO docs.

This helps a lot with matching e.g. Flatpak applications.