Add pref to show summary & body with gauge values
Often it's useful to show the summary and body with a gauge.  For
example, if a volume control is set to above 100%, the gauge will only
go to 100%, but the body text might say something like "Volume: 120%",
which is useful (and possibly very important) information.

However, I can see why many people might not care to see the text values
(I personally do not), so I think it's reasonable to make this a user
preference.