Show summary & body with gauge values, if present
Often it's useful to show the summary and body with a gauge.  For
example, if a volume control is set to above 100%, the gauge will only
go to 100%, but the body text might say something like "Volume: 120%",
which is useful (and possibly very important) information.

In the case where the x-canonical-private-icon-only hint is also
specified, we'll preserve the old behavior.  This is a little janky, as
that hint is really supposed to mean that just the icon should be shown,
but I think it's reasonable to expect that if the value hint is also
sent, the calling app really does want the value shown.