Support Wayland
This uses gtk-layer-shell to place notification windows in the top
layer, so other apps don't obscure the notifications.

As gtk_window_move() doesn't work properly here, we have to use the
layer-shell's anchor and margin settings to place the windows.  We also
adjust the margin settings on the fly when doing the window slideout,
if enabled.