Bump GTK minimum to 2.22 and remove/ifdef X11-isms
This removes some helper functions that check the GTK version, and
allows us to use gdk_monitor_get_workarea(), rather than having to go
through the global list of windows (which isn't accessible on Wayland),
figure out which windows are of type 'dock', and manually compute a
workarea.  (I'm not really sure why this code wasn't just fetching
_NET_WORKAREA from the root window in the first place, but oh well.)

GTK 2.22 was released in 2016, so this should be fine.  People with
distros old enough to be stuck on 2.20 or below probably won't be
upgrading xfce4-notifyd either.