Fix a slide-out loop when the mouse pointer is in the way
If the mouse pointer is sitting in the path of where the notification
will slide off the screen, an enter-notify event gets triggered, which
resets the notification's expiration and position back to the
pre-slide-out spot.

Instead of resetting expiration on enter-notify, now we reset on
motion-notify.  If the pointer is sitting stationary next to the
notification, and the notification moves under it, there won't be any
motion-notify events.

A side-effect of this is that if a new notification appears under the
mouse pointer, it will start expiring immediately.  Previously, it would
be 'held', since an enter-notify event would happen immediately after
the notification appears.  While this is a change in behavior, I'm not
exactly convinced it's bad.

Closes #59.