Add options for what to do when opening the plugin menu
Options are:

* Mark all read
* Mark shown read
* Do nothing

The first one is the default.  Even though it feels more "correct" for
the second one to be the default, I don't want users to get confused
if they leave their computer for a while, get more notifications than
are shown in the menu, open the menu, and then they still have unread
notifications.