Make the log max size a little friendlier
Setting something to '0' to disable it is not intuitive (and is only
documented via a tooltip).  This changes it to be a checkbox, and the
value in the spinbutton now doesn't have any 'special' values.

Initially I was afraid of some data loss if the daemon and settings
dialog are out of sync (during the upgrade from an older version), but I
realized this shouldn't be an issue because if the new settings dialog
is opened while the old daemon is running, the settings dialog will have
already migrated the log to sqlite, and the worst the old daemon can do
is truncate the old keyfile log.