Improve the panel icon when there are unread notifications
Previously we used some semi-custom icon names for when there are new
notifications, but it seems many/most themes don't have this icon.  In
addition, some themes might choose to not put any kind of emblem or
visual aid on the icon when there are new notifications in some states.

So instead, we have our own "emblem", and we draw it on the "default"
icon (either notifiation-symbolic or notification-disabled-symbolic) in
order to ensure we get some kind of indicator on the icon when there are
unread notifications.