Notification windows don't need to be override-redirect
If they are, they can sometimes end up on top of screensaver/screen
locker windows, which is a security issue.

It looks like notification windows were made override-redirect because
some WMs are buggy and will decorate the window without it (see
https://bugzilla.xfce.org/show_bug.cgi?id=10686), but I don't think
that's a good reason to avoid fixing a security issue.  Such WMs should
be fixed, instead.

Closes #7.