Make all notification log access go through dbus
This should fix issues where unread state wasn't getting noticed
properly by the plugin or other entities, and makes log change
notification granular and... well... just better.

Also this allows the daemon to "own" the log and be the only process
that reads from and writes to it.  This way, the plugin and config
dialog can't ever break the DB through version incompatibilities, and
failure modes will be better-handled.

Closes #90.