Remove WM_TAKE_FOCUS from WM_PROTOCOLS for notification windows
GDK unconditionally advertises support for WM_TAKE_FOCUS, even if the
window it's created doesn't accept input focus.  The ICCCM seems to say
that no-input windows should *not* advertise this protocol.

Regardless of whether or not it's correct, xfwm4 will try to give focus
to our notification windows.  Even though that later on fails to happen,
any fullscreen windows will end up being un-fullscreen'd, which is
definitly not what we want to happen.

Even if xfwm4 decides to work around this behavior, I think it makes
sense for xfce4-notifyd to follow the ICCCM properly here.

See xfwm4#705 for more discussion.