Change options for installing startup files
Previously the D-Bus activation files and systemd were not installed by
default, and XDG autostart was the default.  The rationale for using XDG
autostart was to ensure that there's no lag on the first notification,
and also that, if another D-Bus activation service file is installed for
a different DE's notification service, it wouldn't get started instead.

That rationale is reasonable, but can cause problems now that the panel
plugin relies on D-Bus in order to check the notification log: XDG
autostart runs after the panel's startup runs (xfce4-session runs XDG
autostart items after X11 session management apps).

There was also another oddity: systemd unit file installation was tied
to the D-Bus activation file installation, which is probably not quite
right.

So the new defaults (all of which can be changed with configure
switches, which are now all independent):

* D-Bus activation files are installed.
* Systemd unit file is installed.
* XDG autostart file is not installed.

If a user/distro is concerned about notifyd not getting started until
the D-Bus session bus activates it, XDG autostart can be additionally
enabled (and the systemd unit file might actually start it earlier too).

Closes #99.