Instal D-Bus and systemd files under $prefix by default
I thought it might be useful to use pkg-config to find the 'correct'
places to install these files, but I think most people will not expect
files to be installed outside the prefix they select.

The auto-detect behavior can be restored by passing "auto" to the
respective configure arguments.

Addresses comments on #99.