Set initial window position before realizing windows
I don't quite understand why I can't realize the window before calling
gdk_window_move() on it anymore; before 9d2ba01f there was a complicated
ordering to how the window was shown, but that had to change, and that
caused windows to briefly flash wherever the WM wanted to initially
place the window, before it would get moved to the correct location.

It looks like gtk_widget_get_preferred_size() does seem to give the
correct final size, even before realize, so we can use that, it seems,
to place the window ealier.

We still react to the window's size-allocate signal in case things
change later.

Closes #101.