Sanitize instead of validating body text markup
The current code just validates that the body is well-formed XML (well,
the XML subset that GMarkupParser supports).  Instead, we now filter out
tags (and tag attributes) that aren't in the notifications spec.

We also now filter out <img> tags, since we don't support them.  If the
tag has an alt= attribute, we put that text in the notification.

Closes #114