Fix active-monitor notification positioning on Wayland
On Wayland, we can't get the absolute location of the pointer, or even
know which monitor the pointer is on.  Sigh.

Instead, we can create the window with no set monitor, set it's opacity
to fully transparent, map the window, and hope that the compositor's
policy is to place layer-shell windows with no output set on the active
monitor.  Then we can see what monitor the window is on, move it to the
appropriate place, and set the opacity back to normal.

Closes #125.