Remove redundant positioning code from Wayland path
This also changes how XfceNotifyDaemon gets notified that it needs to
reposition the notification, via a new XfceNotifyWindow::need-position
signal, which is only emitted on size-allocate if the new size doesn't
match the last size the daemon class has seen.  This simplifies a bunch
of old code even for the X11 positioning path as well.