Destroy and recreate the panel menu every time it's popped up
Not doing so can cause issues with unnecessary menu scrollbars if the
menu needed scrollbars on one popup, but then on a later popup does not.

I'm not really sure why the menu was being reused each time; I don't
think that really saves much in the way of time since on each popup it
has to clear out the previous contents of the menu anyway.

Not reusing the menu also means that we shouldn't have to call
gtk_menu_reposition() on every size-allocate.

Closes #123.