Add positioning style classes to XfceNotifyWindow
The style classes allow themes to set "direction aware" visual elements
based on the positioning configuration.

In the future, it might make sense to be smarter about setting this
class, based on the actual location the notification ends up being in,
or possibly taking into account the x/y position set in the
notification's hints by the application.

Closes #140