Don't advertise the actions & action-icons capabilities when in DnD
When we're in do-not-disturb mode, notifications will not be shown, and
so the user won't be able to click on action buttons.  For many
notifications this isn't a big deal, but for some (like approving or
rejecting a Bluetooth pairing), the action buttons are essential to the
functionality.

Some apps only seem to check capabilities on startup, so this won't help
them, but blueman, at least, checks more often, and when the
notification daemon doesn't support actions, it will ask for the
approve/reject decision via other means.

Closes #154