build: allow systemd option without systemd installed
systemd.pc is just used to autodetect if the systemd option should be
enabled. If it is manually enabled with -Dsystemd=enabled, just ignore
the systemd dependency requirement and install the service.

Usefull in Alpine Linux where we package systemd files (e.g. for
postmarketOS) but do not ship systemd itself yet.